KA-269 Rename models to .ts
