+++ src/models/ItemViewModel.ts
